face-recognition/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
deepface==0.0.93
Pillow==10.3.0
tf-keras==2.16.0
numpy==1.26.4
